Way too many pushes - mostly Mudlark game
My bad, I ran a lot of things locally here without pushing.

I've now written most of the game, with a couple of intro screens, new functions to detect collisions with the dog (attacking object) and the items (to be found & added to score).
I'll add a ReadME and more information on the website about the development process.

I now need to add functionality for when all the items are found and the game ends, and am also toying with adding a time element too.

If I have time I'd like to tidy up some of the functions as there were a lot of cases where I found a better approach halfway through, and have therefore got slightly inconsistent structures going on. I think there's also a lot of repetition with the hidden items I want to address with loops and objects - but I first wanted to see if I could get it all working.